app2/requirements-mqtt.txt
jsonschema
Flask==1.1.1
pymongo==3.9.0
paho-mqtt==1.2.3
requests==2.21.0
